feat: trlst-216 Allow to remove non-confidential files without a confidential version. (#36)
Normally a public user can only upload a confidential file, followed by the corresponding non-confidential version.
A user managed to upload an extra non-confidential file (possibly by uploading from two different windows).
The extra file is stops him from completing the submission, but there is no provision to remove it.
Added an extra category when showing the uploaded file. The category lists the single non-confidential version, and ask the user to remove it. The interface allows to download the file before deleting it.