fix: trlst 259 issue post tasks
Investigating an issue under TRLST 259 found some non critical issues in caseworker. This PR is a tidy up of the base settings file and the local.env.example file:
- Employed consistent use of django-environ.
- Removed environ.Env.read_env(). We use docker-compose for local development
and feed the container settings from a local.env file. In PaaS we pick up the
app's environment and do not deploy a '.env' file.
- Simplified REDIS URL definition and removed hard coding of database number.
- Removed 'TRUSTED_USER_TOKEN' setting, this is not used in the code base.
- Removed cruft comments.
- Pepified.