feat: Fix countdown and improve page layout.
Fix bug that caused 2FA lock countdown to not be displayed, and improve 2FA challenge layout. This is a quick and dirty fix as the API 2FA views are a mess. If/when the API layer is cleaned up then Public and Caseworker portal's 2FA usage can be refactored to be less abysmal.
- ensure 2fa locked countdown is displayed.
- improve page layout.