Mary's comments in https://uktrade.atlassian.net/browse/TRSV2-589?focusedCommentId=117160
also replaced the file icon retriever with a templatetag so it reverts back to a default if no icon is found for that extension

Replaced the other file icon stuff with the templatetag and also replaced the download icon ticks with a partial.

Reverted changed to design.css and changed the classes in the table manually so they matched the only version. Created custom.css to store custom rules separate from design.css